refactor(buffer): move snapshot/diff cluster onto impl Persistence
Extracts ~290 lines of persistence-owning logic out of impl TextBuffer
and onto impl Persistence, with piece_tree and buffers passed in by
&reference (coordination mechanism b in the plan — borrowed parameters
for read-many paths). Methods moved:

- mark_saved_snapshot(&mut Persistence, &PieceTree)
- refresh_saved_root_if_unmodified(&mut Persistence, &PieceTree)
- apply_chunk_load_to_saved_root(&mut Persistence, ...args)
- diff_since_saved(&Persistence, &PieceTree, &[StringBuffer])
- diff_trees_by_structure(&Persistence, &PieceTree)

Plus private free-fn helpers in persistence.rs:
- tree_total_bytes(&Arc<PieceTreeNode>) -> usize
- extract_range_from_tree(root, start, end, buffers)
- collect_range_from_node(...)
- get_text_range(piece_tree, buffers, offset, bytes)
  (private duplicate of TextBuffer's read-only helper so
  verify_content_differs_in_ranges doesn't reach back into TextBuffer)

TextBuffer retains three thin delegators (mark_saved_snapshot,
refresh_saved_root_if_unmodified, diff_since_saved) — diff_since_saved
keeps the tracing span around the persistence call. One internal
caller of apply_chunk_load_to_saved_root updated to use
self.persistence.apply_chunk_load_to_saved_root.

mod.rs drops from 7,740 → 7,447 lines; persistence.rs grows from
134 → 479.

Full 127-test buffer suite green.
